+++ ruma-client-api/src/r0/capabilities/iter.rs
@@ -1,8 +1,10 @@
+//! Iterator implementation for `Capabilities`
+
 use std::{borrow::Cow, collections::btree_map};
 
 use serde_json::Value as JsonValue;
 
-use super::get_capabilities::Capabilities;
+use super::Capabilities;
 
 /// Reference to a capability.
 #[derive(Debug)]
@@ -13,10 +15,12 @@ pub struct CapabilityRef<'a> {
 }
 
 impl<'a> CapabilityRef<'a> {
+    /// Get name of the capability.
     pub fn name(&self) -> &'a str {
         self.name
     }
 
+    /// Get value of the capability.
     pub fn value(&self) -> Cow<'a, JsonValue> {
         match self.value {
             // unknown capability from btreemap iterator
@@ -33,6 +37,13 @@ pub struct CapabilitiesIter<'a> {
     pub(super) caps: &'a Capabilities,
     pub(super) pos: usize,
     pub(super) custom_caps_iterator: btree_map::Iter<'a, String, JsonValue>,
+}
+
+impl<'a> CapabilitiesIter<'a> {
+    /// Creates a new CapabilitiesIter
+    pub fn new(caps: &'a Capabilities) -> Self {
+        Self { caps, pos: 0, custom_caps_iterator: caps.custom_capabilities.iter() }
+    }
 }
 
 impl<'a> Iterator for CapabilitiesIter<'a> {
